--- classes/main.rb
@@ -1,9 +1,0 @@
-require_relative './app'
-
-def main
-  app = App.new
-
-  app.run
-end
-
-main
